+++ src/isal/igzip.py
@@ -152,9 +152,6 @@ class IGzipFile(gzip.GzipFile):
         s = repr(self.fileobj)
         return '<igzip ' + s[1:-1] + ' ' + hex(id(self)) + '>'
 
-    def flush(self, zlib_mode=isal_zlib.Z_SYNC_FLUSH):
-        super().flush(zlib_mode)
-
     def _write_gzip_header(self, compresslevel=_COMPRESS_LEVEL_TRADEOFF):
         # Python 3.9 added a `compresslevel` parameter to write gzip header.
         # This only determines the value of one extra flag. Because this change
@@ -206,24 +203,18 @@ class _IGzipReader(gzip._GzipReader):
     def __init__(self, fp):
         super().__init__(fp)
         self._decomp_factory = isal_zlib.decompressobj
-        self._decomp_args = dict(wbits=64+isal_zlib.MAX_WBITS)
-        # Set wbits such that ISAL_GZIP_NO_HDR_VER is used. This means that
-        # it does not read a header, and it verifies the trailer.
         self._decompressor = self._decomp_factory(**self._decomp_args)
 
     def _add_read_data(self, data):
-        # isa-l verifies the trailer data, so no need to keep track of the crc.
-        self._stream_size = self._stream_size + len(data)
+        # Use faster isal crc32 calculation and update the stream size in place
+        # compared to CPython gzip
+        self._crc = isal_zlib.crc32(data, self._crc)
+        self._stream_size += len(data)
 
-    def _read_eof(self):
-        # Gzip files can be padded with zeroes and still have archives.
-        # Consume all zero bytes and set the file position to the first
-        # non-zero byte. See http://www.gzip.org/#faq8
-        c = b"\x00"
-        while c == b"\x00":
-            c = self._fp.read(1)
-        if c:
-            self._fp.prepend(c)
+
+# Aliases for improved compatibility with CPython gzip module.
+GzipFile = IGzipFile
+_GzipReader = _IGzipReader
 
 
 # Plagiarized from gzip.py from python's stdlib.